Moves everything to poise.
Easier to just use one framework. Moves event handler to poise. Data
access is now done through the Data struct given to poise.

Also moved things into app. Still not sure how I want to structure
things so just moving things into app as needed.
