+++ src/lib.rs
@@ -1,6 +1,4 @@
 pub mod app;
-pub mod commands;
 pub mod database;
-pub mod handlers;
 pub mod intro;
 pub mod player;
